Final version of libft and modification of printf

+++ CommonCore/bonus_ft_printf/srcs/print_hexalow.c
@@ -10,16 +10,16 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "ft_print.h"
+#include "./../includes/ft_printf.h"
 
 void	print_hexalow(t_print *tab, int fd)
 {
 	char	*value;
-	int	i;
-	
+	int		i;
+
 	value = ft_itoa_base(va_arg(tab->args, unsigned int), "0123456789abcdef");
 	i = -1;
-	while (++i < tab->width - ft_strlen(value))
+	while (++i < tab->width - (int)ft_strlen(value))
 		ft_putchar_fd(' ', fd);
 	ft_putstr_fd(value, fd);
 }
