+++ CommonCore/bonus_ft_printf/srcs/print_integer.c
@@ -10,16 +10,16 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "ft_printf.h"
+#include "./../includes/ft_printf.h"
 
 void	print_integer(t_print *tab, int fd)
 {
 	char	*value;
-	int	i;
+	int		i;
 
 	value = ft_itoa(va_arg(tab->args, int));
 	i = -1;
-	while (++i < tab->width - ft_strlen(value))
+	while (++i < tab->width - (int)ft_strlen(value))
 		ft_putchar_fd(' ', fd);
 	ft_putstr_fd(value, fd);
 }
